move get routes to correct files, update server.js and auth.js to reflect change
