+++ routes/auth.js
@@ -1,16 +1,18 @@
 var authController = require("../controllers/authController.js");
-var htmlRoutes = require("./htmlRoutes");
+var remainingRoutes = require("./remainingRoutes");
+var plannedRoutes = require("./plannedRoutes");
+var spentRoutes = require("./spentRoutes");
 
 module.exports = function(app, passport) {
   app.get("/signup", authController.signup);
 
   app.get("/signin", authController.signin);
 
-  app.get("/user/planned", isLoggedIn, htmlRoutes.userPlanned);
+  app.get("/user/planned", isLoggedIn, plannedRoutes.getUserPlanned);
 
-  app.get("/user/spent", isLoggedIn, htmlRoutes.userSpent);
+  app.get("/user/spent", isLoggedIn, spentRoutes.getUserSpent);
 
-  app.get("/user/remaining", isLoggedIn, htmlRoutes.userRemaining);
+  app.get("/user/remaining", isLoggedIn, remainingRoutes.getUserRemaining);
 
   app.post(
     "/signup",
